axby: switch to partitioning by "a" value
Originally I wanted to create jobs with relatively constant number of
points in the axby space. I tried definining these partitions in terms
of a starting and ending point in the axby space, and also tried
defining them in terms of a starting point and an upper bound on the
number of points.

However, accurately iterating over the space has turned out to be
harder than expected. It would be simplified I think by not using the
iterator interface to the axby space but i want that for easier testing
at this point.

Partitioning by "a" value won't scale to large bases because for large
bases we need to divide the space by both "a" and "b". The first set of
experiments we'll be running aren't going to be huge so doing this keeps
things simple in the short term.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>